Move Tests for User Calls
