+++ tests/protocols/gmpv208/entities/tags/test_delete_tag.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2020-2021 Greenbone Networks GmbH
+# Copyright (C) 2018-2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -16,9 +16,27 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from . import Gmpv214TestCase
-from .testcmds import GmpModifyUserTestCase
+from gvm.errors import GvmError
 
 
-class Gmpv214ModifyUserTestCase(GmpModifyUserTestCase, Gmpv214TestCase):
-    pass
+class GmpDeleteTagTestMixin:
+    def test_delete(self):
+        self.gmp.delete_tag('a1')
+
+        self.connection.send.has_been_called_with(
+            '<delete_tag tag_id="a1" ultimate="0"/>'
+        )
+
+    def test_delete_ultimate(self):
+        self.gmp.delete_tag('a1', ultimate=True)
+
+        self.connection.send.has_been_called_with(
+            '<delete_tag tag_id="a1" ultimate="1"/>'
+        )
+
+    def test_missing_id(self):
+        with self.assertRaises(GvmError):
+            self.gmp.delete_tag(None)
+
+        with self.assertRaises(GvmError):
+            self.gmp.delete_tag('')
